Move functions that extract info about the game state into engine/digest.py
These consist of find_convoys, supplycenters, builds_available, and
actionable_subregions.
